backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pillow==10.4.0
numpy==2.1.1
python-dotenv==1.0.1
httpx==0.27.2
